Refactoring GetByPath trait
